rename package to com.android.chuck
