+++ chuck/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 25
     buildToolsVersion "25.0.2"
     defaultConfig {
-        applicationId "com.android.neuandroid"
+        applicationId "com.android.com.neuandroid.chuck"
         minSdkVersion 14
         targetSdkVersion 25
         versionCode 1
