+++ chuck/src/androidTest/java/com/android/chuck/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.android.neuandroid;
+package com.android.chuck;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("com.android.neuandroid", appContext.getPackageName());
+        assertEquals("com.android.com.neuandroid.chuck", appContext.getPackageName());
     }
 }
