+++ chuck/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.android.neuandroid">
+    package="com.android.chuck">
 
     <uses-permission android:name="android.permission.INTERNET"/>
     <application
@@ -10,7 +10,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name="com.android.chuck.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
