+++ chuck/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.android.neuandroid.MainActivity">
+    tools:context="com.android.chuck.MainActivity">
 
     <GridLayout
         android:layout_width="match_parent"
